feat(scripts): update script name and references
