+++ helpers/.component-template/README.md
@@ -1,3 +1,3 @@
-# @westpac/**COMPONENT_KEY**
+# @westpac/__COMPONENT_KEY__
 
-**COMPONENT_DESCRIPTION**
+__COMPONENT_DESCRIPTION__
